Refactor instellation test
