--- src/janus/data/tests/data_instellation.csv
@@ -1,8 +1,0 @@
-# r [AU],  S_0 [W m-2], OLR [W m-2], net [W m-2], ts [K], tr[K] 
-3.00000e-01,2.34297e+03,1.94397e+03,-4.67185e+01,3.00023e+03,4.19568e+02
-4.83333e-01,9.02641e+02,8.43916e+02,7.70272e+01,2.99961e+03,3.30552e+02
-6.66667e-01,4.74451e+02,5.01731e+02,9.86425e+01,2.99951e+03,2.81454e+02
-8.50000e-01,2.91857e+02,4.21028e+02,1.73069e+02,2.99913e+03,2.49260e+02
-1.03333e+00,1.97482e+02,4.16319e+02,2.48540e+02,2.99876e+03,2.26070e+02
-1.21667e+00,1.42451e+02,4.15915e+02,2.94890e+02,2.99853e+03,2.08342e+02
-1.40000e+00,1.07585e+02,4.15768e+02,3.24365e+02,2.99838e+03,1.94222e+02
