+++ tests/conftest.py
@@ -1,4 +1,5 @@
 import os
+from pathlib import Path
 
 if not os.environ.get("RAD_DIR"):
     raise Exception(
@@ -8,4 +9,3 @@ if not os.environ.get("FWL_DATA"):
     raise Exception(
         "The FWL_DATA environment variable where spectral and evolution tracks data will be downloaded needs to be set up!"
     )
-
